backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
tensorflow-intel==2.13.0
numpy==1.24.3
opencv-python==4.9.0.80
Pillow==10.2.0
python-multipart==0.0.9